Add 'Kute10_54', add 'Kute' base class, global refactor
