+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/evaluation/MUTEExamplesTest.kt
@@ -2,9 +2,12 @@ package com.smeup.rpgparser.evaluation
 
 import com.andreapivetta.kolor.yellow
 import com.smeup.rpgparser.*
-import com.smeup.rpgparser.execution.RunnerCLI.programName
 import com.smeup.rpgparser.mute.color
 import com.smeup.rpgparser.mute.executeMuteAnnotations
+import com.smeup.rpgparser.performance.Kute10_53
+import com.smeup.rpgparser.performance.Kute10_54
+import com.smeup.rpgparser.performance.Kute10_55
+import com.smeup.rpgparser.performance.Kute10_56
 import org.junit.Test
 import org.junit.experimental.categories.Category
 import kotlin.test.assertEquals
@@ -335,6 +338,34 @@ class MUTEExamplesTest {
     // DOU (loop of 10000000 iterations with increment of a numeric var)
     fun execute() {
         assertMuteOK("MUTE10_56")
+    }
+
+    @Test @Category(PerformanceTest::class)
+    // DO (expected 68ms. to perform an empty loop of 10000000 iterations, as MUTE10_53.rpgle)
+    fun executeKute10_53() {
+        val kute = Kute10_53()
+        kute.performanceComparing()
+    }
+
+    @Test @Category(PerformanceTest::class)
+    // FOR (expected 306ms. to perform an empty loop of 10000000 iterations, as MUTE10_54.rpgle)
+    fun executeKute10_54() {
+        val kute = Kute10_54()
+        kute.performanceComparing()
+    }
+
+    @Test @Category(PerformanceTest::class)
+    // DOW (expected 83ms. to perform a loop of 10000000 iterations with increment of a numeric var, as MUTE10_55.rpgle)
+    fun executeKute10_55() {
+        val kute = Kute10_55()
+        kute.performanceComparing()
+    }
+
+    @Test @Category(PerformanceTest::class)
+    // DOU (expected 81ms. to perform a loop of 10000000 iterations with increment of a numeric var, as MUTE10_56.rpgle)
+    fun executeKute10_56() {
+        val kute = Kute10_56()
+        kute.performanceComparing()
     }
 
     private fun dummyProgramFinder() = DummyProgramFinder("/performance/")
